moved trust certificate to different clas so it would only be run once
