+++ collectors/scm/gitlab/src/main/java/com/capitalone/dashboard/collector/GitlabSettings.java
@@ -1,5 +1,15 @@
 package com.capitalone.dashboard.collector;
 
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+
+import javax.annotation.PostConstruct;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.X509TrustManager;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -14,6 +24,9 @@ import org.springframework.stereotype.Component;
 @Component
 @ConfigurationProperties(prefix = "gitlab")
 public class GitlabSettings {
+    
+    private static final Log LOG = LogFactory.getLog(GitlabSettings.class);
+    
     private String cron;
     private String protocol;
     private String host;
@@ -86,5 +99,33 @@ public class GitlabSettings {
 	public void setSelfSignedCertificate(boolean selfSigned) {
 		this.selfSignedCertificate = selfSigned;
 	}
+	
+	@PostConstruct
+    public void trustSelfSignedCertificatesIfNecessary() {
+        if (isSelfSignedCertificate()) {
+            try {
+                final SSLContext ctx = SSLContext.getInstance("TLS");
+                final X509TrustManager tm = new X509TrustManager() {
+                    public void checkClientTrusted(final X509Certificate[] xcs, final String string)
+                            throws CertificateException {
+                    }
+
+                    public void checkServerTrusted(final X509Certificate[] xcs, final String string)
+                            throws CertificateException {
+                    }
+
+                    public X509Certificate[] getAcceptedIssuers() {
+                        X509Certificate[] n = new X509Certificate[0];
+                        return n;
+
+                    }
+                };
+                ctx.init(null, new TrustManager[] { tm }, null);
+                SSLContext.setDefault(ctx);
+            } catch (final Exception ex) {
+                LOG.error(ex.getMessage());
+            }           
+        }
+    }
 
 }
