+++ collectors/scm/gitlab/src/main/java/com/capitalone/dashboard/gitlab/DefaultGitlabGitClient.java
@@ -1,18 +1,10 @@
 package com.capitalone.dashboard.gitlab;
 
 import java.net.URI;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509TrustManager;
-
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -33,8 +25,6 @@ import com.capitalone.dashboard.util.Supplier;
 
 @Component
 public class DefaultGitlabGitClient implements  GitlabGitClient {
-
-    private static final Log LOG = LogFactory.getLog(DefaultGitlabGitClient.class);
 
     //Gitlab max results per page. Reduces amount of network calls.
     private static final int RESULTS_PER_PAGE = 100;
@@ -81,36 +71,9 @@ public class DefaultGitlabGitClient implements  GitlabGitClient {
     }
 
 	private ResponseEntity<GitlabCommit[]> makeRestCall(URI url, String apiToken) {
-		if(gitlabSettings.isSelfSignedCertificate()) {
-			trustSelfSignedSSL();
-		}
 		HttpHeaders headers = new HttpHeaders();
 		headers.add("PRIVATE-TOKEN", apiToken);
 		return restOperations.exchange(url, HttpMethod.GET, new HttpEntity<>(headers), GitlabCommit[].class);
 	}
 
-	private void trustSelfSignedSSL() {
-		try {
-			final SSLContext ctx = SSLContext.getInstance("TLS");
-			final X509TrustManager tm = new X509TrustManager() {
-				public void checkClientTrusted(final X509Certificate[] xcs, final String string)
-						throws CertificateException {
-				}
-
-				public void checkServerTrusted(final X509Certificate[] xcs, final String string)
-						throws CertificateException {
-				}
-
-				public X509Certificate[] getAcceptedIssuers() {
-					X509Certificate[] n = new X509Certificate[0];
-					return n;
-
-				}
-			};
-			ctx.init(null, new TrustManager[] { tm }, null);
-			SSLContext.setDefault(ctx);
-		} catch (final Exception ex) {
-			LOG.error(ex.getMessage());
-		}
-	}
 }
